Added navbar links work
