+++ src/main/resources/templates/fragments/nav.html
@@ -19,18 +19,18 @@
             </div>
             <div class="overflow-hidden px-[5%] lg:flex lg:items-center lg:px-0 lg:[--height-closed:auto] lg:[--height-open:auto]" style="height: var(--height-closed, 0)">
                 <div class="first:pt-4 lg:first:pt-0">
-                    <a href="#" class="block py-3 text-md focus-visible:outline-none lg:px-4 lg:py-2 lg:text-base">Link One</a>
+                    <a href="/home" class="block py-3 text-md focus-visible:outline-none lg:px-4 lg:py-2 lg:text-base">Home</a>
                 </div>
                 <div class="first:pt-4 lg:first:pt-0">
-                    <a href="#" class="block py-3 text-md focus-visible:outline-none lg:px-4 lg:py-2 lg:text-base">Link Two</a>
+                    <a href="/events/new" class="block py-3 text-md focus-visible:outline-none lg:px-4 lg:py-2 lg:text-base">Create event</a>
                 </div>
                 <div class="first:pt-4 lg:first:pt-0">
-                    <a href="#" class="block py-3 text-md focus-visible:outline-none lg:px-4 lg:py-2 lg:text-base">Link Three</a>
+                    <a href="/events" class="block py-3 text-md focus-visible:outline-none lg:px-4 lg:py-2 lg:text-base">Find events</a>
                 </div>
                 <div class="first:pt-4 lg:first:pt-0">
                     <div>
                         <button class="flex w-full items-center justify-between gap-2 py-3 text-left text-md focus-visible:outline-none lg:flex-none lg:justify-start lg:px-4 lg:py-2 lg:text-base">
-                            <span>Link Four</span>
+                            <a href="/user-account" class="block py-3 text-md focus-visible:outline-none lg:px-4 lg:py-2 lg:text-base">Account</a>
                             <span>
                                 <svg stroke="currentColor" fill="none" stroke-width="0" viewBox="0 0 15 15" height="1em" width="1em" xmlns="http://www.w3.org/2000/svg">
                                     <path fill-rule="evenodd" clip-rule="evenodd" d="M3.13523 6.15803C3.3241 5.95657 3.64052 5.94637 3.84197 6.13523L7.5 9.56464L11.158 6.13523C11.3595 5.94637 11.6759 5.95657 11.8648 6.15803C12.0536 6.35949 12.0434 6.67591 11.842 6.86477L7.84197 10.6148C7.64964 10.7951 7.35036 10.7951 7.15803 10.6148L3.15803 6.86477C2.95657 6.67591 2.94637 6.35949 3.13523 6.15803Z" fill="currentColor"></path>
@@ -41,10 +41,10 @@
                 </div>
                 <div class="mt-6 flex flex-col items-center gap-4 lg:ml-4 lg:mt-0 lg:flex-row">
                     <button class="focus-visible:ring-border-primary inline-flex gap-3 items-center justify-center whitespace-nowrap ring-offset-white transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 border border-border-primary text-text-primary bg-background-primary px-5 py-2 w-full">
-                        Button
+                        <a href="/login" class="block py-3 text-md focus-visible:outline-none lg:px-4 lg:py-2 lg:text-base">login</a>
                     </button>
                     <button class="focus-visible:ring-border-primary inline-flex gap-3 items-center justify-center whitespace-nowrap ring-offset-white transition-colors focus-visible:outline-none focus-visible:ring-2 focus-visible:ring-offset-2 disabled:pointer-events-none disabled:opacity-50 border border-border-primary bg-background-alternative text-text-alternative px-5 py-2 w-full">
-                        Button
+                        <a href="/register" class="block py-3 text-md focus-visible:outline-none lg:px-4 lg:py-2 lg:text-base">signup</a>
                     </button>
                 </div>
             </div>
